OcCryptoLib: Renamed files in CamelCase.
