+++ Library/OcCryptoLib/OcCryptoLib.inf
@@ -34,7 +34,7 @@
   Md5.c
   RsaDigitalSign.c
   Sha1.c
-  Sha2-avx.h
+  Sha2Avx.h
   Sha2.c
   SecureMem.c
   PasswordHash.c
@@ -48,8 +48,8 @@
 
 [Sources.X64]
   X64/BigNumWordMul64.c
-  X64/Sha2-avx.c
-  X64/sha512-avx.nasm
+  X64/Sha2Avx.c
+  X64/Sha512Avx.nasm
 
 [FixedPcd]
   gOpenCorePkgTokenSpaceGuid.PcdOcCryptoAllowedRsaModuli
